# docs/swagger.yaml
basePath: /api
definitions:
  controllers.ChangePasswordRequest:
    properties:
      confirm_password:
        example: password_baru123
        type: string
      new_password:
        example: password_baru123
        minLength: 8
        type: string
      old_password:
        example: password_lama123
        type: string
    required:
    - confirm_password
    - new_password
    - old_password
    type: object
  controllers.CreateUserRequest:
    properties:
      jabatan:
        example: ANGGOTA JAGA REGU
        type: string
      kata_sandi:
        example: password123
        minLength: 8
        type: string
      nama_lengkap:
        example: NAMA LENGKAP PETUGAS
        type: string
      nrp:
        example: "98765"
        type: string
      pangkat:
        example: BRIPDA
        type: string
      peran:
        enum:
        - OPERATOR
        - SUPER_ADMIN
        type: string
      regu:
        example: I
        type: string
    required:
    - jabatan
    - kata_sandi
    - nama_lengkap
    - nrp
    - pangkat
    - peran
    type: object
...
  controllers.UpdateProfileRequest:
    properties:
      nama_lengkap:
        example: NAMA SAYA
        type: string
      nrp:
        example: "12345"
        type: string
      pangkat:
        example: BRIPKA
        type: string
    required:
    - nama_lengkap
    - nrp
    - pangkat
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      jabatan:
        description: KANIT SPKT, ANGGOTA JAGA REGU
        type: string
      nama_lengkap:
        type: string
      nrp:
        type: string
      pangkat:
        type: string
      peran:
        description: SUPER_ADMIN, OPERATOR
        type: string
      regu:
        description: I, II, III
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: yusuf.aar@gmail.com
    name: MUHAMMAD YUSUF ABDURROHMAN
    url: https://github.com/dope-s/simdokpol-go
  description: Ini adalah dokumentasi API untuk aplikasi Sistem Informasi Manajemen
    Dokumen Kepolisian.
  license:
    name: MIT License
    url: https://github.com/dope-s/simdokpol-go/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: SIMDOKPOL API
  version: "1.0"
paths:
...
  /profile:
    put:
      consumes:
      - application/json
      description: Memperbarui data profil (Nama, NRP, Pangkat) untuk pengguna yang
        sedang login.
      parameters:
      - description: Data Profil Baru
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Pesan sukses dan data user yang diperbarui
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Terjadi kesalahan pada server'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Memperbarui Profil Pengguna
      tags:
      - Profile
  /profile/password:
    put:
      consumes:
      - application/json
      description: Mengubah kata sandi untuk pengguna yang sedang login.
      parameters:
      - description: Data Kata Sandi Lama dan Baru
        in: body
        name: passwords
        required: true
        schema:
          $ref: '#/definitions/controllers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Pesan Sukses
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 'Error: Kata sandi lama salah'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mengubah Kata Sandi Pengguna
      tags:
      - Profile
...
  /users:
    get:
      description: Mengambil daftar semua pengguna (aktif atau non-aktif). Hanya bisa
        diakses oleh Super Admin.
      parameters:
      - default: active
        description: Filter status pengguna
        enum:
        - active
        - inactive
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: 'Error: Gagal mengambil data pengguna'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mendapatkan Semua Pengguna
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Membuat akun pengguna baru (Operator atau Super Admin). Hanya bisa
        diakses oleh Super Admin.
      parameters:
      - description: Data Pengguna Baru
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: 'Error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Terjadi kesalahan pada server'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Membuat Pengguna Baru
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Masukkan token JWT Anda dengan format 'Bearer {token}'.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"